Corrected Implementation of strategies as per our logic

+++ src/com/hdfclife/strategy/PremiumStrategy.java
@@ -1,5 +1,5 @@
 package com.hdfclife.strategy;
 
 public interface PremiumStrategy{
-    double getPremiumRate();
+    int getPremiumRate();
 }
